@@ -6,4 +6,4 @@
 
 Welcome to your new app.
 
-<MyPrompt Title="How is Blazor working for you?" />
+<MyPrompt Title="Do you like Blazor?" />
